Add `-C sample-profile=FILE` to feed PGO info to LLVM.
This (theoretically) allows LLVM to tune its optimisation decisions
based on the runtime behaviour of the program, as recorded by a
profiler.

At the moment, the only theoretical way to use this that I know is to
profile using some invocation of Linux's `perf` and then convert into
the format LLVM expects with https://github.com/google/autofdo , however
autofdo seems to not work that well: the tool that is meant to convert
to a format compatible with LLVM outputs invalid data (i.e even invalid
if I work with C and clang). (This could be because the exact `perf`
invocation listed in the docs isn't supported on my machine, but using
`perf record -b ...` does mean that `create_llvm_prof` does output
something.)

There are some other downsides/inconsistencies:

- it requires enabling debuginfo for profiling *and* for building the
  PGO'd binary (LLVM segfaults in the latter case). `-C debuginfo=1`
  seems to suffice.

- the memory ownership semantics of adding the pass aren't obvious, so
  the file name is leaked when being passed to LLVM.

- LLVM/rustc's error reporting is horrible. There's theoretically a pile
  of places that will give reasonably detailed errors about an invalid
  profile file, and generally tell the user about problems... but they
  aren't working for me, and I have no idea how the LLVM error reporting
  system works.

- LLVM doesn't seem to make great decisions when given a profile, it
  does change the binary, but it doesn't seem to be in the most useful
  way (the simple example I've been working with got 20% slower). This
  could be caused by me not understanding the format fully, since I'm
  having to write the profiles manually, although it is reasonably
  simple. It could also be caused by our debuginfo/source location info
  not be quite precise enough.

All in all, an interesting experiment, but not usable without more work.